requirements.dev.txt
-r requirements.txt
coverage==5.3
python-dotenv=0.19.2
codecov==2.1.10
Flask-DebugToolbar==0.11.0
flask-debugtoolbar-sqlalchemy==0.2.0
pytest==6.1.1
pytest-cov==2.10.1
Faker==4.14.0
tox==3.20.1
mockredispy==2.9.3